Move auth into config dir
