+++ server/config/auth/index.js
