+++ server/config/auth/logout.js
