+++ server/config/auth/passport_config.js
@@ -1,5 +1,5 @@
-var passport = require('passport');
-const _ = require('lodash');
+const passport = require('passport');
+
 
 // WIP
 passport.serializeUser(function (user, done) {
@@ -12,13 +12,18 @@ passport.deserializeUser(function (user, done) {
 	// placeholder for custom user deserialization.
 	// maybe you are going to get the user from mongo by id?
 	// null is for errors
-	user.de;
 	done(null, user);
 });
 
 
-var GithubStrategy = require('passport-github').Strategy;
-const ghConfig = {
+function responseHandler(req, accessToken, refreshToken, params, profile, done) {
+	return done(null, profile);
+}
+
+const GithubStrategy = require('passport-github').Strategy;
+const GoogleStrategy = require('passport-google-oauth20').Strategy;
+
+const githubConfig = {
 	clientID: '739ab279e272b81baa0b',
 	clientSecret: '1f127e83b84ca87ee5a83445141e1d58cce30979',
 	callbackURL: 'https://localhost:5000/auth/github/callback',
@@ -34,19 +39,7 @@ const googleConfig = {
 	passReqToCallback: true
 };
 
-
-passport.use(new GithubStrategy(ghConfig,
-	function (req, accessToken, refreshToken, params, profile, done) {
-		return done(null, profile);
-	})
-);
-
-var GoogleStrategy = require('passport-google-oauth20').Strategy;
-
-passport.use(new GoogleStrategy(googleConfig,
-	function (req, accessToken, refreshToken, params, profile, done) {
-		return done(null, profile);
-	}
-));
+passport.use(new GithubStrategy(githubConfig,responseHandler));
+passport.use(new GoogleStrategy(googleConfig,responseHandler));
 
 module.exports = passport;
